make the endpoints of follow and unfollow authorized

+++ server/src/controllers/follow.ts
@@ -1,11 +1,19 @@
-import { Request, Response } from "express";
+import { Response } from "express";
+import { CustomRequest } from "../interfaces/CustomRequest";
 import FollowService from "../models/follow";
 
 const followService = new FollowService();
 
 // follow a user function
-const followUser = async (req: Request, res: Response) => {
+const followUser = async (req: CustomRequest, res: Response) => {
 	const { follower_id, followed_id } = req.body;
+	// check if the user logged in
+	if (!req.user) {
+		return res.status(401).json({
+			message: "You must be logged in to follow a user.",
+		});
+	}
+	// check if the user is following the user
 	if (follower_id !== followed_id) {
 		try {
 			const followAUser = await followService.addFollow(
@@ -27,8 +35,16 @@ const followUser = async (req: Request, res: Response) => {
 };
 
 // delete follow function
-const deleteFollow = async (req: Request, res: Response) => {
+const deleteFollow = async (req: CustomRequest, res: Response) => {
 	const { follower_id, followed_id } = req.body;
+
+	// check if the user logged in
+	if (!req.user) {
+		return res.status(401).json({
+			message: "You must be logged in to unfollow a user.",
+		});
+	}
+	// check if the user is following the user
 	if (follower_id !== followed_id) {
 		try {
 			const deleteFollow = await followService.deleteFollow(
